webgis-backend/requirements.txt
flask
flask-cors
gunicorn
sqlalchemy
psycopg2-binary
geopandas
shapely
requests
PyJWT
geoalchemy2
